+++ python-client-sdk/src/sdk/models/operations/authglobal.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Optional
 
 AUTH_GLOBAL_SERVERS = [
 	"http://localhost:35456",
@@ -14,9 +13,9 @@ AUTH_GLOBAL_SERVERS = [
 class AuthGlobalResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""