+++ typescript-client-sdk/src/sdk/models/operations/usageexamplepost.ts
@@ -258,7 +258,7 @@ export class UsageExamplePostResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * A successful response that contains the simpleObject sent in the request body